+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.1.7] - 2018-08-12
+### Added
+- MetaWear.readRSSI - get RSSI value after connection
+- More mocked modules with spoof'd MetaWear's
+
+### Fixed
+- MetaBoot status not updated if scanning wasn't used
+
+## [3.1.6] - 2018-07-31
 ### Added
 - MetaWearScanner.retrieveConnectedMetaWearsAsync function